[refactor] Add code coverage tool; Refactor gem structure to proper one; Leave only one constant in foreign `Rack` module
