+++ Gemfile
@@ -8,3 +8,6 @@ group :test do
   gem "webmock"
 end
 
+group :development, :test do
+  gem "simplecov"
+end
