+++ lib/rack_reverse_proxy/errors.rb
@@ -1,0 +1,33 @@
+module RackReverseProxy
+  module Errors
+    class GenericURI < Exception
+      attr_reader :url
+
+      def intialize(url)
+        @url = url
+      end
+
+      def to_s
+        %Q(Your URL "#{@url}" is too generic. Did you mean "http://#{@url}"?)
+      end
+    end
+
+    class AmbiguousMatch < Exception
+      attr_reader :path, :matches
+      def initialize(path, matches)
+        @path = path
+        @matches = matches
+      end
+
+      def to_s
+        %Q(Path "#{path}" matched multiple endpoints: #{formatted_matches})
+      end
+
+      private
+
+      def formatted_matches
+        matches.map {|matcher| matcher.to_s}.join(', ')
+      end
+    end
+  end
+end
