+++ lib/rack_reverse_proxy/version.rb
@@ -1,0 +1,3 @@
+module RackReverseProxy
+  VERSION = "0.9.1"
+end
